Add typings for manage_buy_offer.
